% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DJV.R
\name{Binning_DJV}
\alias{Binning_DJV}
\title{Loads a dataframe for binning}
\usage{
Binning_DJV(data, bin = 5)
}
\arguments{
\item{data}{Input dataframe}
}
\value{
A dataframe
}
\description{
This function is used for binning 
Bins the variables and retuns a data frame
}
