% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DJV.R
\name{Correlation_DJV}
\alias{Correlation_DJV}
\title{Load a list of correlated variables}
\usage{
Correlation_DJV(Correlated_variables, data)
}
\arguments{
\item{Correlated_variables}{list of correlated variables}
}
\value{
best variable value
}
\description{
It loads the function to find the best variables
}
