% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DJV.R
\name{Feeder_High_School_Function}
\alias{Feeder_High_School_Function}
\title{Feeder High school cutoff selection}
\usage{
Feeder_High_School_Function(Model_data)
}
\arguments{
\item{data}{Input dataframe}
}
\value{
A dataframe
}
\description{
This function is used for Feeder High school cutoff
and best cutoff is returned as dataframe
}
